+++ 2D_Graphics_Project/Group_Project/Polyline.cpp
@@ -76,13 +76,13 @@ void Polyline::Move(int x, int y) override
     Draw();
 }
 
-int Polyline::Perimeter() override
+double Polyline::Perimeter() override
 {
     // has no perimeter dumby function 
     return -1; // error value returned
 }
 
-float Polyline::Area() override
+double Polyline::Area() override
 {
     // has no area dumby function 
     return -1; // error value returned
