+++ 2D_Graphics_Project/Group_Project/Rectangle.h
@@ -21,8 +21,8 @@ class Rectangle: public Shape
 
         void Draw() override;
         void Move(int x, int y) override;
-        int Perimeter() override;
-        float Area() override;
+        double Perimeter() override;
+        double Area() override;
 
     private:
         int height;
