+++ 2D_Graphics_Project/Group_Project/Text.h
@@ -23,8 +23,8 @@ class Text: public Shape
 
         void Draw() override;
         void Move(int x, int y) override;
-        int Perimeter() override;
-        float Area() override;
+        double Perimeter() override;
+        double Area() override;
 
     private:
         string textString;
